Add view-based Emoji search fronted
