+++ gradle/libs.versions.toml
@@ -60,5 +60,6 @@ zipline-loader = { module = "app.cash.zipline:zipline-loader", version.ref = "zi
 paparazzi-gradlePlugin = { module = "app.cash.paparazzi:paparazzi-gradle-plugin", version = "1.1.0" }
 jimfs = "com.google.jimfs:jimfs:1.2"
 coil-compose = { module = "io.coil-kt:coil-compose", version = "2.2.2" }
+coil-core = { module = "io.coil-kt:coil", version = "2.2.2" }
 
 lint-core = { module = "com.android.tools.lint:lint", version.ref = "lint" }
