+++ samples/emoji-search/android-composeui/src/main/AndroidManifest.xml
