+++ samples/emoji-search/android-composeui/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
