+++ samples/emoji-search/android-composeui/src/main/res/mipmap-xxxhdpi/ic_launcher.webp
