--- samples/emoji-search/android/src/main/java/app/cash/zipline/samples/emojisearch/ComposeUiText.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright (C) 2022 Square, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package app.cash.zipline.samples.emojisearch
-
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-import app.cash.redwood.LayoutModifier
-import example.schema.widget.Text
-
-class ComposeUiText : Text<@Composable () -> Unit> {
-  private var text by mutableStateOf("")
-
-  override var layoutModifiers: LayoutModifier = LayoutModifier
-
-  override val value = @Composable {
-    Text(
-      text = text,
-      color = MaterialTheme.colors.onBackground,
-    )
-  }
-
-  override fun text(text: String) {
-    this.text = text
-  }
-}
