+++ settings.gradle
@@ -40,7 +40,8 @@ include ':samples:counter:sunspot:compose:protocol'
 include ':samples:counter:sunspot:widget'
 include ':samples:counter:sunspot:widget:protocol'
 
-include ':samples:emoji-search:android'
+include ':samples:emoji-search:android-composeui'
+include ':samples:emoji-search:android-views'
 include ':samples:emoji-search:ios:shared'
 include ':samples:emoji-search:launcher'
 include ':samples:emoji-search:presenters'
